+++ lib/new_relic/agent/error_trace_aggregator.rb
@@ -81,9 +81,10 @@ module NewRelic
           # Already seen this class once? Bail!
           return if @errors.any? { |err| err.exception_class_name == exception.class.name }
 
-          trace = exception.backtrace || caller.dup
-          noticed_error = NewRelic::NoticedError.new('NewRelic/AgentError', exception)
-          noticed_error.stack_trace = trace
+          noticed_error = NewRelic::Agent.instance.error_collector.create_noticed_error(exception,
+            {metric: 'NewRelic/AgentError'})
+          noticed_error.stack_trace = caller.dup unless exception.backtrace
+
           @errors << noticed_error
         end
       rescue => e